refactor(core): address transaction-related review feedback
