+++ packages/core/src/index.ts
@@ -32,4 +32,3 @@ export * from './naming-strategy';
 export * from './metadata';
 export * from './cache';
 export * from './decorators';
-export * from './transaction';
